backend/requirements.txt
asgiref==3.11.0
cross-web==0.4.0
Django==6.0
django-choices-field==4.0.0
graphql-core==3.2.7
packaging==25.0
pillow==12.1.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
strawberry-graphql[django]>=0.254.0
django-cors-headers>=4.6.0
typing_extensions==4.15.0
